% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_GermVersity}
\alias{run_GermVersity}
\title{Run the Shiny Application}
\usage{
run_GermVersity(...)
}
\arguments{
\item{...}{arguments to pass to golem_opts.
See `?golem::get_golem_options` for more details.}
}
\description{
Run the Shiny Application
}
